requirements.txt
praw
psaw
ipykernel
tqdm
mysql-connector-python 
